Add ROStests for aws_file_uploader

+++ aws_uploader/src/aws_uploader/aws_file_uploader.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import logging
 import boto3
 import os
@@ -18,10 +20,42 @@ class AwsFileUploader:
         self._create_bucket_if_doesnt_exist(default_bucket, self._default_region)
         self._default_bucket = default_bucket
 
+    # Note: overrides all files with the same name
+    def upload_directory_contents_to_aws_directory(
+            self,
+            directory_path,
+            bucket_name=None,
+            is_remove_file_on_success=True,
+    ):
+
+        if os.path.abspath(directory_path) == os.getcwd():
+            raise IOError("Directory selected cannot be directory of this file: '{}'".format(directory_path))
+        if self._is_parent_directory(directory_path):
+            raise IOError("Directory selected cannot have other directories in it: '{}'".format(directory_path))
+
+        file_names = os.listdir(directory_path)
+        for file_name in file_names:
+            file_path = os.path.join(directory_path, file_name)
+            self.upload_file(
+                file_path,
+                bucket_name=bucket_name,
+                is_remove_file_on_success=is_remove_file_on_success,
+            )
+
+    @staticmethod
+    def _is_parent_directory(path):
+        for item in os.listdir(path):
+            if os.path.isdir(os.path.join(path, item)):
+                return True
+        return False
+
     def upload_file(self, file_path, bucket_name=None, is_remove_file_on_success=False):
 
         if bucket_name is None:
             bucket_name = self._default_bucket
+
+        if not os.path.exists(file_path):
+            raise ReferenceError("No file '{}' exists".format(file_path))
 
         try:
             self._s3_client.upload_file(file_path, bucket_name, os.path.basename(file_path))
@@ -94,9 +128,9 @@ if __name__ == '__main__':
     if not os.path.exists(temp_dir):
         os.mkdir(temp_dir)
 
-    file_path = os.path.join(temp_dir, temp_file)
-    with open(file_path, 'w') as f:
+    file_path_ = os.path.join(temp_dir, temp_file)
+    with open(file_path_, 'w') as f:
         f.write("Some content generated at {}".format(datetime.datetime.now().strftime("%H:%M:%S on %Y-%m-%d")))
 
-    aws_uploader.upload_file(file_path, is_remove_file_on_success=True)
+    aws_uploader.upload_file(file_path_, is_remove_file_on_success=True)
     os.rmdir(temp_dir)
